# backend-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de consultation du stock pour l'application mobile Android (POC)",
        "title": "Stock API — WMS SPEED",
        "contact": {
            "name": "BK Systèmes",
            "email": "jerome.gouez@bksystemes.fr"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/activities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne toutes les activités actives (ACT_ACTF=1) de la table ACT_PAR",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activities"
                ],
                "summary": "Liste des activités",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ActivityListResponse"
                        }
                    },
                    "401": {
                        "description": "Token manquant ou invalide",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ActivityListResponse"
                        }
                    }
                }
            }
        },
        "/api/health/api": {
            "get": {
                "description": "Retourne 200 si le processus API est actif (ne nécessite pas SQL Server)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Santé de l'API",
                "responses": {
                    "200": {
                        "description": "status=healthy",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/health/database": {
            "get": {
                "description": "Teste la connexion à SQL Server et retourne la version et l'heure serveur",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Connexion SQL Server",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Authentifie un utilisateur et retourne un JWT valable 24h",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authentification",
                "parameters": [
                    {
                        "description": "Identifiants",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login réussi — contient le token",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "hash_method invalide",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Login ou mot de passe incorrect",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "500": {
                        "description": "Erreur base de données ou JWT",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/stock/details/{sku}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne les informations détaillées d'un article (non encore implémenté)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Détail stock par SKU",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code article (SKU)",
                        "name": "sku",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Token manquant ou invalide",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/stock/search": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Recherche des articles en stock filtrés par activité. Au moins un critère parmi art_code, stk_lieu, stk_nosu est obligatoire.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "Recherche stock",
                "parameters": [
                    {
                        "description": "Critères de recherche",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StockSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats (items peut être vide)",
                        "schema": {
                            "$ref": "#/definitions/models.StockSearchResponse"
                        }
                    },
                    "401": {
                        "description": "Token manquant ou invalide",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Corps invalide ou act_code manquant",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ActivityItem": {
            "type": "object",
            "properties": {
                "act_code": {
                    "type": "string"
                },
                "act_keyu": {
                    "type": "integer"
                },
                "act_lib": {
                    "type": "string"
                }
            }
        },
        "models.ActivityListResponse": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ActivityItem"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string"
                }
            }
        },
        "models.HealthResponse": {
            "type": "object",
            "properties": {
                "database_status": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "models.LoginRequest": {
            "type": "object",
            "properties": {
                "hash_method": {
                    "description": "\"CLAIR\", \"MD5\", or \"SHA256\"",
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "description": "Only on success (86400 for 24h)",
                    "type": "integer"
                },
                "log_location": {
                    "description": "Only on error",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string"
                },
                "token": {
                    "description": "Only on success",
                    "type": "string"
                }
            }
        },
        "models.StockItem": {
            "type": "object",
            "properties": {
                "art_code": {
                    "type": "string"
                },
                "qua_code": {
                    "type": "string"
                },
                "stk_lieu": {
                    "type": "string"
                },
                "stk_nosu": {
                    "type": "string"
                },
                "stk_qte": {
                    "type": "integer"
                }
            }
        },
        "models.StockSearchRequest": {
            "type": "object",
            "properties": {
                "act_code": {
                    "description": "Activity code (required, filtered from Android)",
                    "type": "string"
                },
                "art_code": {
                    "type": "string"
                },
                "stk_lieu": {
                    "type": "string"
                },
                "stk_nosu": {
                    "type": "string"
                }
            }
        },
        "models.StockSearchResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StockItem"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT obtenu via POST /api/login. Format : \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}